cleaned up refactored the tools and agent
