+++ innovation_pathfinder_ai/source_container/container.py
@@ -1,0 +1,1 @@
+all_sources = []
